+++ src/lib.rs
@@ -8,6 +8,7 @@ mod program;
 mod shader;
 mod stack;
 mod texture;
+mod vertices;
 #[cfg(target_arch = "wasm32")]
 mod webgl;
 #[cfg(test)]
@@ -26,3 +27,15 @@ pub use crate::shader::Shader;
 pub use crate::texture::Builder as TextureBuilder;
 pub use crate::texture::Texture;
 pub use crate::texture::TextureInfo;
+pub use crate::vertices::VertexArray;
+pub use crate::vertices::VertexBuffer;
+
+pub mod vertex {
+  pub use crate::vertices::Attrib;
+  pub use crate::vertices::AttribType;
+  pub use crate::vertices::Attribs;
+  pub use crate::vertices::VertexP3f as P3f;
+  pub use crate::vertices::VertexP3fN3f as P3fN3f;
+  pub use crate::vertices::VertexP3fT2f as P3fT2f;
+  pub use crate::vertices::VertexP3fT2fN3f as P3fT2fN3f;
+}