+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Unreleased
 - Introduced `MatrixStack` type
 - Introduced `Shader` and `Program` types
 - Introduced `Texture` type and supporting infrastructure
+- Introduced `VertexArray` and `VertexBuffer` types and associated
+  functionality
 
 
 0.1.0